feat: Add tax calculator page and translations
