+++ src/pages/Home/__tests__/index.spec.tsx
@@ -1,6 +1,5 @@
 import { screen, render } from '@testing-library/react'
 import '@testing-library/jest-dom'
-
 import Home from '../index.tsx'
 
 describe('Home', () => {
